Updating codes
